+++ rasa/core/featurizers.py
@@ -13,7 +13,7 @@ from rasa.core.domain import PREV_PREFIX, Domain
 from rasa.core.events import ActionExecuted
 from rasa.core.trackers import DialogueStateTracker
 from rasa.core.training.data import DialogueTrainingData
-from rasa.utils.common import disable_logging
+from rasa.utils.common import is_logging_disabled
 
 logger = logging.getLogger(__name__)
 
@@ -472,7 +472,7 @@ class FullDialogueTrackerFeaturizer(TrackerFeaturizer):
             "collected trackers (by {}({}))..."
             "".format(type(self).__name__, type(self.state_featurizer).__name__)
         )
-        pbar = tqdm(trackers, desc="Processed trackers", disable=disable_logging())
+        pbar = tqdm(trackers, desc="Processed trackers", disable=is_logging_disabled())
         for tracker in pbar:
             states = self._create_states(tracker, domain, is_binary_training=True)
 
@@ -580,7 +580,7 @@ class MaxHistoryTrackerFeaturizer(TrackerFeaturizer):
             "collected trackers (by {}({}))..."
             "".format(type(self).__name__, type(self.state_featurizer).__name__)
         )
-        pbar = tqdm(trackers, desc="Processed trackers", disable=disable_logging())
+        pbar = tqdm(trackers, desc="Processed trackers", disable=is_logging_disabled())
         for tracker in pbar:
             states = self._create_states(tracker, domain, True)